Added new Features: Find all references, highlight usages, and rename refactoring
Also reverted to single instance of the F# mime type until the old one
is removed from MonoDevelop.  Changing the mime type affects the
parsing of files because the first mime type registered for the .fs
extension takes precedence and the ones we define here in the binding
don’t take effect.