Set minimum threads to 8 for Mono 4.2.1
This was causing deadlocks on Mono as the default
number of threads was insufficient. The library function

    System.Threading.ThreadPool.SetMinThreads

also did not work.

This has been worked around in Emacs by setting the
environment variable MONO_THREADS_PER_CPU to 8:

    https://github.com/fsharp/emacs-fsharp-mode/pull/46

This has been reported and fixed in Mono:

    https://bugzilla.xamarin.com/show_bug.cgi?id=37288

So now the call to `SetMinThreads` should be sufficient.

Fixes #88.