Fixed issue with array notation
Weirdly
```
let test = []
```
and
```
let test : int array = []
```
Both produce FSharEntities where the first IsArray and the second is not, so you have to unannotate the second, api is such a mess in FCS.

Second part of this commit simply does QuoteIdentifierIfNeeded only on FSharpTypes that have a TypeDefinition, anything else uses the Format.  Im not exactly sure what doesn have a TypeDefinition apart froma  generic parameter which is covered in the other branch.