Fix how we assign FSI options
For netcore/netfx we should always explicitly provide the targetprofile,
because the impacts how reference resolution works (ie which primary
assembly we look for and key off of), and we should also provide any
static additional arguments as the 'otherFlags' argument.

In addition, we need to explicitly not provide an mscorlib reference
when targeting netstandard, as that re-enables some of the legacy paths.

Finally, there was a bug in the handling of the 'replace refs' function
that was overwriting all of the OtherOptions instead of modifying just
the framework refs.