Fix template list parsing on non-English systems
To list the available templates, FSAC parses directly the output of `dotnet new --list`, looking for predefined English words. But if the configured display language / locale of the system is another language (e.g. French or German), this parsing fails and generates an out-of-bounds exception.

This commit takes the simplest route to fix the problem: forcing `dotnet` to print its output in English by setting a documented environment variable[1].

Fixes ionide/ionide-vscode-fsharp#1569

[1] https://docs.microsoft.com/en-us/dotnet/core/tools/dotnet-environment-variables#dotnet_cli_ui_language