Many tweaks to AdaptiveLSPServer for performance (#1024)
* Minor tweaks to adaptiveserver for performance

* Updates FileSystem GetLastWriteTimeShim

This updates GetLastWriteTimeShim to use actualFs.GetLastWriteTimeShim which should help with how the type checker recognizes doing caching for type checks. Additionally updates DidOpen, DidChange, and DidSave to use the file systems GetLastWriteTimeUtc to further help with this type check caching.

* Fixes the .BufferedDebounce to actualy buffer

this was processing observable windows as they came in instead of waiting for them the close.