Update README.md
Hello. This adds Flow and Kate, two editors, to the list of apps that are using FSAutoComplete. 

I added the respective project websites as the attached link, since they are using it without special configuration or installation. 

That seems to make the homepage of those comparable unknown editors more useful, with the link to the commit that added support. 

For confirmation, can you still see the commits here:

https://github.com/neurocyte/flow/blob/a299d27b1fbe0109903eb3b0e0fa942cb0ec18a1/src/file_type_lsp.zig#L51-L53

https://invent.kde.org/utilities/kate/-/blob/c2792fc2f988b10cf038c8d05f5d4e115a5ff899/addons/lspclient/settings.json

https://github.com/helix-editor/helix/blob/master/languages.toml#L52

I will try to get the superfluous flag in the Kate configuration removed.

Helix seems to use `AutomaticWorkspaceInit` by default. 
I assume this is correct.