Test Explorer Support via VSTest (#1383)
* Set up the stub of a test explorer test project

* Add the test explorer tests to the sln

* Set up a trival test case to scaffold for VSTestAdapter work

* Update Expecto for performance improvements

* Demonstrate the most basic test discovery with vstest

* Demonstrate connection between ionide and language server

* Implement and mvp test discovery endpoint on the Language Server

* Push test discovery down to AdaptiveState

This is preparation for adding notifications for incremental discovery reporting

* Publish incremental test discovery updates

* Demonstration a test run using VSTest

* Rename the namespace for test features.

VSTestAdapter is too specific.
We already know we'll need to expand to accomodate Microsoft.Testing.Platform

* Add an lsp endpoint for running tests

* Fix bug where the TestExplorer tests don't build the required sample projects

* Demonstrate basic streaming of test run results from langauge server

* Add test filtering to the VSTest wrapper

* Add testCaseFilter support to test Runs

Still has a bug where NUnit doesn't respect the filter, but works for all other frameworks

* Prove that I can scrape test host process Ids

Also required figuring out graceful cancellation behavior for test runs

* Convert VsTestWrapper methods to async-only for graceful cancellation

* Pass a server-level test for the attaching debugger to test run.

* Format files

* Delete some diagnostic code

* fix incorrect message type on the runTests lsp message

* Pass unit tests using RunTestsWithCustomTestHost

RunTestsWithCustomTestHost is the official way to attach a debugger with VSTest

* Debug test runs

Broke out the debug attach message because it needs to be a blocking call.
VSTest will continue executing the tests as soon as AttachDebuggerToProcess returns

I also fixed the LSP level test explorer tests, which were failing because the sample test projects needed to be built at part of the unit test, since the build artifacts are being cleared before every test run

* Test that client environment variables are available in a test run

* Decouple SampleTestProjects for Lsp and TestExplorer test projects

FsAutoComplete.Tests.TestExplorer expects the sample projects to be built once before the tests are run,
but FsAutoComplete.Tests.Lsp expects the sample projects to be build for every test.
Splitting up the sample projects reduces chances that they could break each other when the run in parallel or in random order

* Forward test logs

* Allow client to limit which test projects are run

This was mainly motivated by test debugging.

Before this feature, the debugger would launch for every test project even if you only wanted to debug one test.
With this change, it'll only launch the one project.

This also partially mitigates the bug where NUnit doesn't respect filters, now those test won't show unless some tests from an nunit project are selected

* Workaround for NUnit not respecting test filters

NUnit doesn't respect test filters when VSTest is in DesignMode, which it is by default with VsTestConsoleWrapper

https://github.com/ionide/FsAutoComplete/pull/1383#issuecomment-3245590606

* Forward test discovery logs to the client for improved error diagnostics

* Show error notification in client for expectable test discovery issues (i.e projects not restored or built)

* Separate discovery vs run lsp-level test explorer tests

* Split the VSTest wrapper code from TestServer contracts for clarity

* Clarify dictionary name (projectLookup -> projectsByBinaryPath)

* Satisfy fantomas formatting check

* Remove print statements per copilot review

* Add OpenTelemetry to the Test Explorer langauge server endpoints

* Avoid duplicate project load when discovery test projects

* Fix net9 build

---------

Co-authored-by: Jimmy Byrd <jimmy.byrd@motivity.net>