updated readme to reflect what's currently implemented
some of the readme was out of date with what is actually implemented and
used, so I thought I'd update the readme to reflect that.

I had it done by copilot cli, so if there are any mistakes I'd be happy
to correct them, though I think it's right, from what I can see.